requirements.txt
git+https://github.com/jungin500/py-hanspell
git+https://github.com/haven-jeon/PyKoSpacing.git
openai==0.27.8
yake==0.4.8
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
bs4==0.0.1
django-allauth==0.54.0
python-docx==0.8.11
torch==2.0.1
transformers==4.30.2
django-widget-tweaks==1.4.12
bardapi==0.1.18
django-extensions==3.2.3
keybert==0.7.0       
django-cors-headers==4.1.0
python-mecab-ko==1.3.3 